Changed way alternate keybinds work.
This change made it so that each key can have a list of keys that can be used instead of the original. This prevents the following situation that was possible with old system:

Keybind = Left Control + Z
AltKeys = Right Control

If left and right control are pressed, it fires.

Also I am getting rid of emojis in commit messages.